# perf: reduce import overhead on background jobs