Avoids 7MB of overhead and cleanup costs for each background job.


Memory profile of worker horse doing `frappe.ping`

![image](https://github.com/frappe/frappe/assets/9079960/44936e55-5025-45d7-b16c-1a8e53c7c346)



After

![image](https://github.com/frappe/frappe/assets/9079960/09107448-e111-405e-8afd-9dc27b3e53f6)


Also, several other modules are loaded instantly on new jobs. In worker pool we can save on this CPU cost by preloading (paying shared memory cost). This increases memory usage of 8 workers by 20MB but reduces import overheads on each job by almost 60-70%

![image](https://github.com/frappe/frappe/assets/9079960/15ccc562-3608-4b33-b049-1053637f7cfb)




---

performing 100 job (`frappe.ping`) - good enough proxy for overheads. 


Before:
```
λ bench worker-pool --num-workers=8 --burst > /dev/null  
8.67s user 5.18s system 491% cpu 2.815 total
```

After:
```
λ time bench worker-pool --num-workers=8 --burst > /dev/null
1.71s user 2.56s system 179% cpu 2.376 total
```


performing 100 job (`frappe.email.queue.flush`) - realistic work

Before:
```
λ time bench worker-pool --num-workers=8 --burst > /dev/null
26.98s user 11.11s system 597% cpu 6.378 total
```

After:
```
λ time bench worker-pool --num-workers=8 --burst > /dev/null
8.61s user 4.73s system 510% cpu 2.613 total
```