@ankush 2024-03-15
Analyzing typical imports that we can save for worker-pool:


Email queue:
- safe_exec - 6mb 
- requests - 5mb
- PIL - 1.5mb
- filetype - 1.4mb
- SQLParse - 2mb